Add files via upload
Smart City Tourist Guide is a web-based software developed in Java programming language to store the details of a city. The main purpose of this project to help the tourists
Safer communication and improved infrastructure. 
A smart city solution for tourists has significant potential to improve the tourist experience, enhance the city's infrastructure and services, and drive economic growth and sustainability.